requirements.txt
discord.py>=1.6.0
requests>=2.22.0
Pillow>=8.1.0
pyocr==0.8